feat: added upload functionality
Reformatted packages into separate files
Added auxiliary package for generating random string
